feat: update astro
