+++ .astro/settings.json
@@ -1,5 +1,5 @@
 {
 	"_variables": {
-		"lastUpdateCheck": 1719317528442
+		"lastUpdateCheck": 1741537064754
 	}
 }
